requirements.txt
pytorch
torchvision
gpytorch
scikit-learn
scikit-image
numpy
matplotlib
seaborn
tqdm